requirements.txt
# Core Dependencies
torch>=2.0.0
torch-geometric>=2.3.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
networkx>=3.0
scikit-learn>=1.3.0

# Data Processing
pyshp>=2.3.1
shapely>=2.0.0
geopandas>=0.13.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0

# Graph Partitioning Baselines
metis>=0.2a5
python-louvain>=0.16
